@@ -1,10 +1,10 @@
 name: textfield_state
 description: A helper class for maintaining a TextEditingController and/or a FocusNode for a State and its Widgets.
-version: 0.2.1
+version: 0.3.0
 homepage: https://github.com/defy-logic/textfield_state
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
