@@ -19,12 +19,12 @@ class TextFieldState {
   /// They are provided here for convenience instead of having to call
   /// `update()` after construction.
   TextFieldState({
-    _TextChanged textChanged,
-    _FocusChanged focusChanged,
-    _FocusChanged primaryFocusChanged,
-    String text,
-    TextEditingController controller,
-    FocusNode focusNode,
+    _TextChanged? textChanged,
+    _FocusChanged? focusChanged,
+    _FocusChanged? primaryFocusChanged,
+    String? text,
+    TextEditingController? controller,
+    FocusNode? focusNode,
   })  : textChanged = textChanged,
         focusChanged = focusChanged,
         primaryFocusChanged = primaryFocusChanged {
@@ -32,82 +32,80 @@ class TextFieldState {
     _updateFocusNode(focusNode);
   }
 
-  final _TextChanged textChanged;
-  final _FocusChanged focusChanged;
-  final _FocusChanged primaryFocusChanged;
+  final _TextChanged? textChanged;
+  final _FocusChanged? focusChanged;
+  final _FocusChanged? primaryFocusChanged;
 
-  bool get _usesFocusNode =>
-      focusChanged != null || primaryFocusChanged != null;
+  bool get _usesFocusNode => focusChanged != null || primaryFocusChanged != null;
 
-  TextEditingController get controller => _widgetController ?? _controller;
-  TextEditingController _widgetController;
-  TextEditingController _controller;
+  TextEditingController get controller => _widgetController ?? _controller!;
+  TextEditingController? _widgetController;
+  TextEditingController? _controller;
 
-  FocusNode get focusNode => _widgetFocusNode ?? _focusNode;
-  FocusNode _widgetFocusNode;
-  FocusNode _focusNode;
+  FocusNode get focusNode => _widgetFocusNode ?? _focusNode!;
+  FocusNode? _widgetFocusNode;
+  FocusNode? _focusNode;
 
-  String _widgetInitialText;
-  String _prevText;
-  bool _hadFocus;
-  bool _hadPrimaryFocus;
+  String? _widgetInitialText;
+  String? _prevText;
+  bool _hadFocus = false;
+  bool _hadPrimaryFocus = false;
 
-  String get text => controller?.text ?? '';
+  String get text => controller.text;
 
-  void _updateController(TextEditingController widgetController, String text) {
+  void _updateController(TextEditingController? widgetController, String? text) {
     if (widgetController != _widgetController) {
       final hasController = widgetController != null;
       final hadController = _widgetController != null;
 
       if (hadController) {
-        _widgetController.removeListener(_handleControllerChanged);
+        _widgetController?.removeListener(_handleControllerChanged);
 
         if (!hasController) {
           assert(_controller == null);
-          if (text != null &&
-              text != _widgetInitialText &&
-              text != _widgetController.text) {
+          if (text != null && text != _widgetInitialText && text != _widgetController?.text) {
             _controller = TextEditingController(text: text);
           } else {
-            _controller =
-                TextEditingController.fromValue(_widgetController.value);
+            _controller = TextEditingController.fromValue(_widgetController?.value);
           }
-          _controller.addListener(_handleControllerChanged);
+          _controller!.addListener(_handleControllerChanged);
         }
       }
 
       if (hasController) {
-        final defaultText = text ?? widgetController.text;
+        final defaultText = text ?? widgetController?.text;
 
         /// Don't reset text if reconstructing with the same initial value.
         if (defaultText == _widgetInitialText) {
-          widgetController.value = controller.value;
+          if (widgetController != null) {
+            widgetController.value = controller.value;
+          }
         } else {
           _widgetInitialText = defaultText;
         }
         _controller?.dispose();
         _controller = null;
-        widgetController.addListener(_handleControllerChanged);
+        widgetController?.addListener(_handleControllerChanged);
       }
     }
 
     _widgetController = widgetController;
 
     if (_controller != null) {
-      final dirty = text != null &&
-          text != _widgetInitialText &&
-          text != _controller.text;
+      final dirty = text != null && text != _widgetInitialText && text != _controller?.text;
 
       _widgetInitialText = text;
-      if (dirty) _controller.text = text;
+      if (dirty) {
+        _controller!.text = text!;
+      }
     } else {
       // Ensure initialized
       _controller = TextEditingController(text: text ?? _prevText ?? '');
-      _controller.addListener(_handleControllerChanged);
+      _controller!.addListener(_handleControllerChanged);
     }
   }
 
-  void _updateFocusNode(FocusNode widgetFocusNode) {
+  void _updateFocusNode(FocusNode? widgetFocusNode) {
     if (!_usesFocusNode) return;
 
     if (widgetFocusNode != _widgetFocusNode) {
@@ -123,9 +121,9 @@ class TextFieldState {
     _widgetFocusNode = widgetFocusNode;
 
     // Ensure initialized
-    if (focusNode == null) {
+    if (_widgetFocusNode == null && _focusNode == null) {
       _focusNode = FocusNode();
-      _focusNode.addListener(_handleFocusNodeChanged);
+      _focusNode!.addListener(_handleFocusNodeChanged);
     }
   }
 
@@ -138,25 +136,21 @@ class TextFieldState {
   /// If [text] is passed, it will become the (given or current) controller's
   /// text unless it matches the controller's initial value.
   void update({
-    TextEditingController controller,
-    FocusNode focusNode,
-    String text,
+    TextEditingController? controller,
+    FocusNode? focusNode,
+    String? text,
   }) {
     _updateController(controller, text);
     _updateFocusNode(focusNode);
   }
 
+  // TODO this is a bad idea
   /// Call this during [State.dispose()].
   void dispose() {
     _controller?.dispose();
-    _controller = null;
     _focusNode?.dispose();
-    _focusNode = null;
     _widgetController?.removeListener(_handleControllerChanged);
     _widgetFocusNode?.removeListener(_handleFocusNodeChanged);
-    _prevText = null;
-    _hadFocus = null;
-    _hadPrimaryFocus = null;
   }
 
   void _handleControllerChanged() {
